requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
jinja2

boto3
python-dotenv
httpx

os
geoip2
motor
pyyaml

llama-index 
llama-index-llms-ollama 
llama-index-embeddings-huggingface
